Add to cart and doing Cart Screen reload

+++ lib/representation/widgets/item_check_out_widget.dart
@@ -12,7 +12,7 @@ class CheckoutItemWidget extends StatelessWidget {
   });
 
   final String? name;
-  final double? price;
+  final int? price;
   final int? quantity;
   final String? quantitative;
